(#415) Elegant solution to a hard problem
Basically, the code check in each act if there are changes to the NEXT act.

If there are, the only way this could happen is through an abandoned data from an erased save file at the same used save slot.

So, it sets the Act value of that change to Integer.MAX_VALUE, where it won't EVER be used. Then sooner or later that run will end and all the data involving that will be erased.